Add domain package and move UseCases into it
