+++ app/src/main/java/com/defaultapps/moviebase/di/component/ApplicationComponent.java
@@ -5,12 +5,12 @@ import android.support.annotation.Nullable;
 import com.defaultapps.moviebase.App;
 import com.defaultapps.moviebase.data.firebase.FavoritesManager;
 import com.defaultapps.moviebase.data.local.AppPreferencesManager;
-import com.defaultapps.moviebase.data.usecase.DiscoverUseCase;
-import com.defaultapps.moviebase.data.usecase.GenreUseCase;
-import com.defaultapps.moviebase.data.usecase.HomeUseCase;
-import com.defaultapps.moviebase.data.usecase.MovieUseCase;
-import com.defaultapps.moviebase.data.usecase.PersonUseCase;
-import com.defaultapps.moviebase.data.usecase.SearchUseCase;
+import com.defaultapps.moviebase.domain.usecase.DiscoverUseCase;
+import com.defaultapps.moviebase.domain.usecase.GenreUseCase;
+import com.defaultapps.moviebase.domain.usecase.HomeUseCase;
+import com.defaultapps.moviebase.domain.usecase.MovieUseCase;
+import com.defaultapps.moviebase.domain.usecase.PersonUseCase;
+import com.defaultapps.moviebase.domain.usecase.SearchUseCase;
 import com.defaultapps.moviebase.di.module.AnalyticsModule;
 import com.defaultapps.moviebase.di.module.ApplicationModule;
 import com.defaultapps.moviebase.di.module.RepositoryModule;
