+++ app/src/main/java/com/defaultapps/moviebase/di/module/UseCaseModule.java
@@ -1,17 +1,17 @@
 package com.defaultapps.moviebase.di.module;
 
-import com.defaultapps.moviebase.data.usecase.DiscoverUseCase;
-import com.defaultapps.moviebase.data.usecase.DiscoverUseCaseImpl;
-import com.defaultapps.moviebase.data.usecase.GenreUseCase;
-import com.defaultapps.moviebase.data.usecase.GenreUseCaseImpl;
-import com.defaultapps.moviebase.data.usecase.HomeUseCase;
-import com.defaultapps.moviebase.data.usecase.HomeUseCaseImpl;
-import com.defaultapps.moviebase.data.usecase.MovieUseCase;
-import com.defaultapps.moviebase.data.usecase.MovieUseCaseImpl;
-import com.defaultapps.moviebase.data.usecase.PersonUseCase;
-import com.defaultapps.moviebase.data.usecase.PersonUseCaseImpl;
-import com.defaultapps.moviebase.data.usecase.SearchUseCase;
-import com.defaultapps.moviebase.data.usecase.SearchUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.DiscoverUseCase;
+import com.defaultapps.moviebase.domain.usecase.DiscoverUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.GenreUseCase;
+import com.defaultapps.moviebase.domain.usecase.GenreUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.HomeUseCase;
+import com.defaultapps.moviebase.domain.usecase.HomeUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.MovieUseCase;
+import com.defaultapps.moviebase.domain.usecase.MovieUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.PersonUseCase;
+import com.defaultapps.moviebase.domain.usecase.PersonUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.SearchUseCase;
+import com.defaultapps.moviebase.domain.usecase.SearchUseCaseImpl;
 
 import dagger.Binds;
 import dagger.Module;
