+++ app/src/main/java/com/defaultapps/moviebase/domain/base/UseCase.java
@@ -1,4 +1,4 @@
-package com.defaultapps.moviebase.data.base;
+package com.defaultapps.moviebase.domain.base;
 
 public interface UseCase {
     void dispose();
