+++ app/src/main/java/com/defaultapps/moviebase/ui/base/BasePresenter.java
@@ -6,7 +6,7 @@ import easybind.PresenterClass;
 import easybind.calls.OnAttach;
 import easybind.calls.OnDetach;
 import easybind.calls.OnDispose;
-import com.defaultapps.moviebase.data.base.UseCase;
+import com.defaultapps.moviebase.domain.base.UseCase;
 
 import io.reactivex.disposables.CompositeDisposable;
