+++ app/src/main/java/com/defaultapps/moviebase/ui/person/PersonPresenterImpl.java
@@ -1,6 +1,6 @@
 package com.defaultapps.moviebase.ui.person;
 
-import com.defaultapps.moviebase.data.usecase.PersonUseCase;
+import com.defaultapps.moviebase.domain.usecase.PersonUseCase;
 import com.defaultapps.moviebase.di.scope.PerFragment;
 import com.defaultapps.moviebase.ui.base.BasePresenter;
