+++ app/src/test/java/com/defaultapps/moviebase/domain/usecase/GenreUseCaseTest.java
@@ -1,4 +1,4 @@
-package com.defaultapps.moviebase.data.usecase;
+package com.defaultapps.moviebase.domain.usecase;
 
 import com.defaultapps.moviebase.data.TestSchedulerProvider;
 import com.defaultapps.moviebase.data.local.AppPreferencesManager;
