+++ app/src/test/java/com/defaultapps/moviebase/domain/usecase/SearchUseCaseTest.java
@@ -1,8 +1,8 @@
-package com.defaultapps.moviebase.data.usecase;
+package com.defaultapps.moviebase.domain.usecase;
 
 
 import com.defaultapps.moviebase.data.models.responses.movies.MoviesResponse;
-import com.defaultapps.moviebase.data.repository.ApiRepository;
+import com.defaultapps.moviebase.domain.repository.ApiRepository;
 import junit.framework.Assert;
 
 import org.junit.Before;
