+++ app/src/test/java/com/defaultapps/moviebase/ui/discover/DiscoverPresenterTest.java
@@ -2,7 +2,7 @@ package com.defaultapps.moviebase.ui.discover;
 
 
 import com.defaultapps.moviebase.data.models.responses.genres.Genres;
-import com.defaultapps.moviebase.data.usecase.DiscoverUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.DiscoverUseCaseImpl;
 
 import org.junit.Before;
 import org.junit.Test;
