+++ app/src/test/java/com/defaultapps/moviebase/ui/genre/GenrePresenterTest.java
@@ -1,7 +1,7 @@
 package com.defaultapps.moviebase.ui.genre;
 
 import com.defaultapps.moviebase.data.models.responses.movies.MoviesResponse;
-import com.defaultapps.moviebase.data.usecase.GenreUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.GenreUseCaseImpl;
 
 import org.junit.Before;
 import org.junit.Test;
