+++ app/src/test/java/com/defaultapps/moviebase/ui/home/HomePresenterTest.java
@@ -1,7 +1,7 @@
 package com.defaultapps.moviebase.ui.home;
 
 import com.defaultapps.moviebase.data.models.responses.movies.MoviesResponse;
-import com.defaultapps.moviebase.data.usecase.HomeUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.HomeUseCaseImpl;
 import com.defaultapps.moviebase.utils.AppConstants;
 import com.defaultapps.moviebase.utils.TrampolineBusThreadScheduler;
 import com.defaultapps.moviebase.utils.rx.RxBus;
