+++ app/src/test/java/com/defaultapps/moviebase/ui/search/SearchPresenterTest.java
@@ -2,7 +2,7 @@ package com.defaultapps.moviebase.ui.search;
 
 
 import com.defaultapps.moviebase.data.models.responses.movies.MoviesResponse;
-import com.defaultapps.moviebase.data.usecase.SearchUseCaseImpl;
+import com.defaultapps.moviebase.domain.usecase.SearchUseCaseImpl;
 
 import org.junit.Before;
 import org.junit.Test;
